requirements.txt
Django==1.6.1
argparse==1.2.1
django-browserid==0.9
djangorestframework==2.3.10
fancy-tag==0.2.0
requests==2.1.0
six==1.4.1
wsgiref==0.1.2
